% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicPrice.R
\name{basicPrice}
\alias{basicPrice}
\title{Basic Price Histogram}
\usage{
basicPrice(url, node)
}
\arguments{
\item{url}{url from shopping website}

\item{node}{data structure that contains price and item}
}
\value{
basic histogram of prices obtained from node
}
\description{
Basic Price Histogram
}
\examples{
\dontrun{
basicPrice(url, node)
}
}
