% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currencyFunc.R
\name{currencyFunc}
\alias{currencyFunc}
\title{Currency Changing Function to Peso, Euro, or USD}
\usage{
currencyFunc(euro = 0.89 * usd, usd = 1.13 * euro, peso = 21 * usd)
}
\arguments{
\item{euro}{type of currency}

\item{usd}{type of currency}

\item{peso}{type of currency}
}
\value{
list of all values and corresponding currency
}
\description{
Currency Changing Function to Peso, Euro, or USD
}
\examples{
\dontrun{
currencyFunc(usd= 2.78)
currencyFunc(peso= 1200)
}
}
