% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priceAna.R
\name{priceAna}
\alias{priceAna}
\title{Histogram of means for shopping item Categories}
\usage{
priceAna(url, node)
}
\arguments{
\item{url}{url from shopping website}

\item{node}{data structure that contains price and item}
}
\value{
histogram of means for all the types of items obtained in node
}
\description{
Histogram of means for shopping item Categories
}
\examples{

priceAna("https://www.thereformation.com/categories/new?gclid=CjwKCAiAm7OMBhAQEiwArvGi3EYt2jozZbzQktEYy2iInclcxHpJKKnWfqu3Jo2mHfABJ-R90ghQVhoCmwAQAvD_BwE" , ".product-prices__price span , .product-summary__name a")
}
