% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priceNum.R
\name{priceNum}
\alias{priceNum}
\title{Takes prices from node and parses it and makes it a numerical}
\usage{
priceNum(html, node)
}
\arguments{
\item{html}{html code for website}

\item{node}{data scruture that contains prices}
}
\value{
parsed numerical values
}
\description{
Takes prices from node and parses it and makes it a numerical
}
\examples{
\dontrun{
priceNum(resultsZara,".price__amount-current")}
}
