% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samel.R
\name{samel}
\alias{samel}
\title{Making two lists the same length by repeating values}
\usage{
samel(x, y)
}
\arguments{
\item{x}{first list of items}

\item{y}{second list of item}
}
\value{
finds longest list and makes other list the same length by repeating values
}
\description{
Making two lists the same length by repeating values
}
\examples{
\dontrun{
samel(x, y)
samel(list1, list2)
}

}
