% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samelNA.R
\name{samelNA}
\alias{samelNA}
\title{Making two lists the same length by adding NA's}
\usage{
samelNA(x, y)
}
\arguments{
\item{x}{first list of items}

\item{y}{second list of item}
}
\value{
finds longest list and makes other list the same length with added NA's
}
\description{
Making two lists the same length by adding NA's
}
\examples{
\dontrun{
samelNA(rev, zara)
samelNa(list1, list2)
}
}
