% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shopScrape.R
\name{shopScrape}
\alias{shopScrape}
\title{Web scrapping shopping sites and creating interactive data frames}
\usage{
shopScrape(url, node, url2, node2)
}
\arguments{
\item{url}{first url from shopping website}

\item{node}{first data structure that contains price and item}

\item{url2}{second url from shopping website}

\item{node2}{second data structure that contains price and item}
}
\value{
data frame
}
\description{
Web scrapping shopping sites and creating interactive data frames
}
\examples{
\dontrun{
shopScrape(url, node, url2, node2)
}
shopScrape("https://www.thereformation.com/categories/new?gclid=CjwKCAiAm7OMBhAQEiwArvGi3EYt2jozZbzQktEYy2iInclcxHpJKKnWfqu3Jo2mHfABJ-R90ghQVhoCmwAQAvD_BwE" , ".product-prices__price span , .product-summary__name a","https://www.zara.com/us/en/woman-must-haves-l4503.html?v1=1609213", ".price__amount-current , #main .link span")
}
